requirements.txt
diffusers==0.21.4
einops==0.8.0
facenet_pytorch==2.5.3
ipython==8.12.3
kornia==0.6.8
matplotlib==3.7.2
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.16.1
pandas==2.2.2
Pillow==8.4.0
pip==23.3
python-dotenv==1.0.1
pytorch_lightning==1.9.0
Requests==2.31.0
scikit_learn==1.3.2
scipy==1.13.0
setuptools==68.0.0
six==1.16.0
tensorboard==2.16.2
tensorboardX==2.6.2.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.1
transformers==4.34.1
wandb==0.16.0
git+https://github.com/Phoveran/fastargs.git@main#egg=fastargs
git+https://github.com/CompVis/taming-transformers.git@master#egg=taming-transformers
git+https://github.com/openai/CLIP.git@main#egg=clip